+++ src/constants/tokens.constant.ts
@@ -1,0 +1,5 @@
+import { MINUTE } from './times.constant';
+
+export const ACCESS_TOKEN_DURATION = MINUTE;
+
+export const REFRESH_TOKEN_DURATION = 5 * MINUTE;