+++ src/modules/jwt-manager/jwt-manager.types.ts
@@ -1,5 +1,6 @@
 export interface UserAccessTokenPayload {
     login: string;
+    expirationTimestamp: number;
     capabilities?: {
         canAdd?: boolean;
         canEdit?: boolean;
@@ -8,6 +9,11 @@ export interface UserAccessTokenPayload {
     isAdmin?: boolean;
 }
 
+export type UserAccessTokenPayloadDto = Omit<UserAccessTokenPayload, 'expirationTimestamp'>;
+
 export interface UserRefreshTokenPayload {
     login: string;
+    expirationTimestamp: number;
 }
+
+export type UserRefreshTokenPayloadDto = Omit<UserRefreshTokenPayload, 'expirationTimestamp'>;