+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.23.0] - 2024-02-14
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Refresh token endpoint at `/users/refreshTokens`.
+- [Marek Kurańda](https://github.com/mjkuranda): JwtManager utils.
+
 ## [1.22.0] - 2024-02-12
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `MealService/getMeals`.